atualizando estrutura de arquivos scripts
